# openapi/openapi.json
{
  "components": {
    "schemas": {
      "AcceptedOrder": {
        "additionalProperties": false,
        "description": "Minimal 202 response; Provider and correlation facts stay private.",
        "properties": {
          "order_id": {
            "format": "uuid",
            "title": "Order Id",
            "type": "string"
          },
          "status": {
            "pattern": "^QUEUED$",
            "title": "Status",
            "type": "string"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          }
        },
        "required": [
          "order_id",
          "status",
          "status_url"
        ],
        "title": "AcceptedOrder",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PartnerInviteAcceptRequest": {
        "additionalProperties": false,
        "properties": {
          "token": {
            "maxLength": 128,
            "minLength": 32,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "PartnerInviteAcceptRequest",
        "type": "object"
      },
      "PartnerInviteConsentRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "order_intent_hash": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Order Intent Hash",
            "type": "string"
          },
          "order_intent_id": {
            "format": "uuid",
            "title": "Order Intent Id",
            "type": "string"
          },
          "partner_asset_id": {
            "format": "uuid",
            "title": "Partner Asset Id",
            "type": "string"
          }
        },
        "required": [
          "expected_version",
          "order_intent_id",
          "order_intent_hash",
          "partner_asset_id"
        ],
        "title": "PartnerInviteConsentRequest",
        "type": "object"
      },
      "PartnerInviteCreateRead": {
        "additionalProperties": false,
        "properties": {
          "invite": {
            "$ref": "#/components/schemas/PartnerInviteSnapshot"
          },
          "join_url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Join Url",
            "type": "string"
          },
          "token": {
            "maxLength": 128,
            "minLength": 32,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "invite",
          "token",
          "join_url"
        ],
        "title": "PartnerInviteCreateRead",
        "type": "object"
      },
      "PartnerInviteCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "template_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Template Id",
            "type": "string"
          }
        },
        "required": [
          "template_id"
        ],
        "title": "PartnerInviteCreateRequest",
        "type": "object"
      },
      "PartnerInviteOrderRequest": {
        "additionalProperties": false,
        "properties": {
          "consent_event_id": {
            "format": "uuid",
            "title": "Consent Event Id",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "host_asset_id": {
            "format": "uuid",
            "title": "Host Asset Id",
            "type": "string"
          }
        },
        "required": [
          "expected_version",
          "host_asset_id",
          "consent_event_id"
        ],
        "title": "PartnerInviteOrderRequest",
        "type": "object"
      },
      "PartnerInviteSnapshot": {
        "additionalProperties": false,
        "properties": {
          "consent_event_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Event Id"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "intent_policy_version": {
            "title": "Intent Policy Version",
            "type": "string"
          },
          "order_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "order_intent_hash": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Order Intent Hash",
            "type": "string"
          },
          "order_intent_id": {
            "format": "uuid",
            "title": "Order Intent Id",
            "type": "string"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "template_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Template Id",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "purpose",
          "status",
          "role",
          "version",
          "expires_at",
          "order_intent_id",
          "order_intent_hash",
          "intent_policy_version",
          "template_id"
        ],
        "title": "PartnerInviteSnapshot",
        "type": "object"
      },
      "PartnerInviteVersionRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "PartnerInviteVersionRequest",
        "type": "object"
      },
      "PartnerWithdrawalRead": {
        "additionalProperties": false,
        "properties": {
          "case_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "case_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Status"
          },
          "invite_id": {
            "format": "uuid",
            "title": "Invite Id",
            "type": "string"
          },
          "invite_status": {
            "title": "Invite Status",
            "type": "string"
          },
          "order_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Status"
          }
        },
        "required": [
          "invite_id",
          "invite_status"
        ],
        "title": "PartnerWithdrawalRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API for AI Wedding Photo generation using InstantID",
    "title": "AI Wedding Photo API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/partner-invites": {
      "post": {
        "operationId": "create_invite_api_v1_partner_invites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerInviteCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerInviteCreateRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Invite",
        "tags": [
          "partner-invites"
        ]
      }
    },
    "/api/v1/partner-invites/accept": {
      "post": {
        "operationId": "accept_invite_api_v1_partner_invites_accept_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerInviteAcceptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerInviteSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Invite",
        "tags": [
          "partner-invites"
        ]
      }
    },
    "/api/v1/partner-invites/{invite_id}": {
      "get": {
        "operationId": "get_invite_api_v1_partner_invites__invite_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "invite_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invite Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerInviteSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Invite",
        "tags": [
          "partner-invites"
        ]
      }
    },
    "/api/v1/partner-invites/{invite_id}/consent": {
      "post": {
        "operationId": "record_consent_api_v1_partner_invites__invite_id__consent_post",
        "parameters": [
          {
            "in": "path",
            "name": "invite_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invite Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerInviteConsentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerInviteSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Consent",
        "tags": [
          "partner-invites"
        ]
      }
    },
    "/api/v1/partner-invites/{invite_id}/order": {
      "post": {
        "operationId": "create_invite_order_api_v1_partner_invites__invite_id__order_post",
        "parameters": [
          {
            "in": "path",
            "name": "invite_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invite Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerInviteOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedOrder"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Invite Order",
        "tags": [
          "partner-invites"
        ]
      }
    },
    "/api/v1/partner-invites/{invite_id}/revoke": {
      "post": {
        "operationId": "revoke_invite_api_v1_partner_invites__invite_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "invite_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invite Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerInviteVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerInviteSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Invite",
        "tags": [
          "partner-invites"
        ]
      }
    },
    "/api/v1/partner-invites/{invite_id}/withdraw": {
      "post": {
        "operationId": "withdraw_invite_api_v1_partner_invites__invite_id__withdraw_post",
        "parameters": [
          {
            "in": "path",
            "name": "invite_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invite Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerInviteVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerWithdrawalRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Withdraw Invite",
        "tags": [
          "partner-invites"
        ]
      }
    },
...
  }
}